tidy_kpathsea: start this
This is a clone of kpathsea that I will gradually tidy up. That way it will be
easier for me to understand how the rest of TeX actually integrates into
kpathsea and what features it uses, in preparation for swapping it out with
different backend that works with Zip files.