otangle: more work on getting named constants out
It turns out that there are places where they are involved in arithmetic
expressions where we can't express the constant's role as valid Pascal. So, we
instead are taking a tack where we embed the derivation of the numerical
constant as a comment. Current version compiles to C and generates a valid
executable. Unfortunately, something in web2c is discarding the comments.