Start work on bringing underscores back.
We just need to stop having otangle get rid of them! Then we need to update a
bunch of the .defines files to match the new symbols.

But for the emitted C code to actuall compile, we will have to alias a bunch
of things that were implemented in C with missing underscores.