Update source code to TeXLive 45838
This is commit b9b7d9db0541ede69b97cdbeba778599a931c919 of the Preining
mirror. This isn't any particular tag or branch or anything, it's just the
most recent checkout I have at the moment (2017 Nov 17), because I want to
validate that the framework is still working and see how much has changed
since I last synced up with upstream.

(The Preining mirror doesn't currently provide tags or branches, so
unfortunately I can't use it to pin the code to a specific release like
TeXLive 2017.1. Hopefully that will change in the future.)